fix: move Encoder manipulator do Api
